+++ iivs/common/visualization/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ("auto_rescale",)
+
+from iivs.common.visualization.contrast import auto_rescale